+++ lib/tasks/version/pre_alpha.ex
@@ -14,7 +14,7 @@ defmodule Mix.Tasks.Version.Alpha do
     |> Config.create()
     |> Git.is_clean()
     |> bump_alpha()
-    |> Changelog.update()
+    |> Changelog.pre_release_update()
     |> Version.update_mix_file()
     |> Git.tag_with_new_version()
     |> Hex.publish()